Break mocking out into separate folder
